+++ libraries/helpers.rb
@@ -2,8 +2,11 @@
 # Cookbook:: chrony
 # Libraries:: helpers
 #
-# Author: Dang H. Nguyen <dang.nguyen@disney.com>
+# Author:: Dang H. Nguyen <dang.nguyen@disney.com>
+# Contributor:: Lance Albertson <lance@osuosl.org>
+#
 # Copyright:: 2020, The Walt Disney Company, All Rights Reserved
+# Copyright:: 2020, Sous Chefs
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -16,33 +19,26 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-#
+module Chrony
+  module Cookbook
+    module Helpers
+      def chrony_service_name
+        if platform_family?('rhel', 'fedora', 'amazon')
+          'chronyd'
+        else
+          'chrony'
+        end
+      end
 
-# this systemd? helper method might collide with Chef Infra Client >= 15.5
-# leaving here for older Chef Infra Clients
-unless respond_to?(:systemd?)
-  def systemd?
-    ::File.exist?('/proc/1/comm') && ::File.new('/proc/1/comm').gets.chomp == 'systemd'
+      def chrony_conf_file
+        if platform_family?('rhel', 'fedora', 'amazon')
+          '/etc/chrony.conf'
+        else
+          '/etc/chrony/chrony.conf'
+        end
+      end
+    end
   end
 end
-
-def chrony_service_path
-  value_for_platform_family(
-    %w(rhel fedora amazon) => '/usr/lib/systemd/system/chronyd.service',
-    'default' => '/lib/systemd/system/chrony.service'
-  )
-end
-
-def chrony_service_name
-  value_for_platform_family(
-    %w(rhel fedora amazon) => 'chronyd',
-    'default' => 'chrony'
-  )
-end
-
-def chrony_conf_file
-  value_for_platform_family(
-    %w(rhel fedora amazon) => '/etc/chrony.conf',
-    'default' => '/etc/chrony/chrony.conf'
-  )
-end
+Chef::DSL::Recipe.include Chrony::Cookbook::Helpers
+Chef::Resource.include Chrony::Cookbook::Helpers