+++ recipes/master.rb
@@ -1,8 +1,10 @@
 #
 # Author:: Matt Ray <matt@@chef.io>
+# Contributor:: Lance Albertson <lance@osuosl.org>
 # Cookbook:: chrony
 # Recipe:: master
 # Copyright:: 2011-2020, Chef Software, Inc.
+# Copyright:: 2020, Sous Chefs
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -19,20 +21,9 @@
 
 package 'chrony'
 
-systemd_unit "#{chrony_service_name}.service" do
-  action %i(create enable)
-  content node['chrony']['systemd']
-  verify false
-  only_if { systemd? }
-end
-
 service 'chrony' do
   service_name chrony_service_name
   supports restart: true, status: true, reload: true
-  if systemd? && docker?
-    start_command "systemctl --no-block start #{chrony_service_name}"
-    restart_command "systemctl --no-block restart #{chrony_service_name}"
-  end
   action %i(start enable)
 end
 
@@ -40,23 +31,6 @@ end
 # ip = node.ipaddress.split('.')
 # set the allowed hosts to the class B
 # node['chrony'][:allow] = ["allow #{ip[0]}.#{ip[1]}"]
-
-# if there are NTP servers, use the first 3 for the initslew
-if node['chrony']['servers'].empty?
-  clients = search(:node, 'recipes:chrony\:\:client').sort || []
-  unless clients.empty?
-    node.default['chrony']['initslewstep'] = 'initslewstep 10'
-    count = 3
-    count = clients.length if clients.length < count
-    count.times { |x| node.default['chrony']['initslewstep'] += " #{clients[x].ipaddress}" }
-  end
-else
-  node.default['chrony']['initslewstep'] = 'initslewstep 10'
-  keys = node['chrony']['servers'].keys.sort
-  count = 3
-  count = keys.length if keys.length < count
-  count.times { |x| node.default['chrony']['initslewstep'] += " #{keys[x]}" }
-end
 
 template 'chrony.conf' do
   path chrony_conf_file