+++ test/cookbooks/test/metadata.rb
@@ -1,0 +1,6 @@
+name              'test'
+maintainer        'Sous Chefs'
+maintainer_email  'help@sous-chefs.org'
+license           'Apache-2.0'
+description       'Testing cookbook for chrony'
+version           '0.0.1'