+++ test/cookbooks/test/recipes/default.rb
@@ -1,0 +1,3 @@
+# Workaround issue running chronyd in docker related to:
+# https://access.redhat.com/solutions/4410831
+directory '/var/run/chrony' if platform_family?('rhel')