+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 This file is used to list changes made in each version of the chrony cookbook.
 
+## Unreleased
+
+- Sous Chefs Adoption
+- Standardize helpers library
+- Improve testing
+- Remove unused attributes
+- Create attribute to handle custom extra configuration settings
+
 ## 0.3.1 (2020-12-23)
 
 - Minor updates to the kitchen files and metadata.rb source fix - [@tas50](https://github.com/tas50)